requirements.txt
telethon
aiohttp
python-dotenv
loguru
flake8
pytest